requirements.txt
fastapi==0.111.0
langchain==0.1.17
langchain_anthropic==0.1.11
langchain_community==0.0.37
langchain_core==0.1.52
langchain_openai==0.1.6
langchain_text_splitters==0.0.1
pydantic==2.7.1
python-dotenv==1.0.1
SQLAlchemy==2.0.29
uvicorn==0.29.0
pymysql==1.0.2
faiss-cpu
firebase-admin
google-auth
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
pinecone
pytest